realizar middleware para los productos. Close #14

+++ src/controller/products.js
@@ -4,13 +4,18 @@ const createProduct = async (req, res) => {
   const {
     name, price, image, type,
   } = req.body;
+
+  if (!name || !price || typeof price !== 'number') {
+    // return next(400);
+    return res.status(400).json('indicar nombre o precio');
+  }
   const newProduct = new Product({
     name, price, image, type,
   });
 
   const productSave = await newProduct.save();
   // productSave.then((e) => res.json(e));
-  res.status(201).json(productSave);
+  res.status(200).json(productSave);
   // console.log(productSave);
   // res.json(req.body);
 };
@@ -21,20 +26,33 @@ const getProduct = async (req, res) => {
 };
 
 const getProductById = async (req, res) => {
+  /* validar que el producto exista */
   const product = await Product.findById(req.params.productId);
   if (!product) return res.status(404).json('product id not found in database');
+
   return res.status(200).json(product);
 };
 
 const updateProductById = async (req, res) => {
+  /* validar que el producto exista */
+  const product = await Product.findById(req.params.productId);
+  if (!product) return res.status(404).json('product id not found in database');
+
+  /* validar que se indique propiedad a modificar */
+  if (Object.entries(req.body).length === 0 || typeof req.body.price !== 'number') return res.status(400).json('no ingreso dato a actualizar');
+
   const updateProduct = await Product.findByIdAndUpdate(req.params.productId, req.body,
     { new: true });
   res.status(200).json(updateProduct);
 };
 
 const deleteProductById = async (req, res) => {
+  /* validar que el producto exista */
+  const product = await Product.findById(req.params.productId);
+  if (!product) return res.status(404).json('product id not found in database');
+
   await Product.findByIdAndDelete(req.params.productId);
-  res.status(204).json();
+  res.status(200).json(product);
 };
 
 module.exports = {
